feat: Adds routing structure
