+++ package.json
@@ -24,6 +24,7 @@
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-hook-form": "^7.62.0",
+    "react-router": "^7.9.1",
     "tailwind-variants": "^1.0.0",
     "tailwindcss": "^4.1.10",
     "zod": "^3.22.4"
