+++ src/pages/page-home.tsx
@@ -1,0 +1,9 @@
+import Text from "../components/text";
+
+export default function PageHome() {
+    return(
+        <>
+            <Text variant="heading-medium">Página inicial!</Text>
+        </>
+    )
+}
